# ⚡️Optimize loading of children in default pagination