This should have been done in #230 but I missed it. The default pagination behavior should have the optimization that only loads the kf_ids of child entities. This functionality was only being executed if a study_id query param was supplied. I moved it into the default query for pagination.