@dankolbman 2018-04-20
Any chance you tried benchmarking these to see how much difference it made?