@znatty22 2018-04-20
Haven't tried benchmarking but when I have sqlalchemy echo on, I see that without this optimization, its makes an additional query for every child entity of a parent. I don't think we ever want that if its possible to get all of the data in one query.