@znatty22 2018-04-20
Also just realized I believe you actually did benchmark this improvement in #213!