@dankolbman 2018-04-20
Yes, was curious about how it performed for the different entities.