@znatty22 2018-04-20
Hmm, you got me curious too :). I ran the perf tests for each of the entities that have child entities. 
I ran the tests on all of the data we have loaded so far (4 studies - Seidman, Schiffman, Chung, Rios)

### Baseline - Participants
```
wrk -t1 -c12 -d60s 'http://localhost:5000/participants?limit=100'                                                                                                                  singhn4@8c85907178da
Running 1m test @ http://localhost:5000/participants?limit=100
  1 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.21s     0.00us   1.21s   100.00%
    Req/Sec     0.00      0.00     0.00    100.00%
  52 requests in 1.00m, 4.57MB read
  Socket errors: connect 0, read 0, write 0, timeout 51
Requests/sec:      0.87
Transfer/sec:     77.95KB
```
### Optimized - Participants
```
wrk -t1 -c12 -d60s 'http://localhost:5000/participants?limit=100'                                                                                                                  singhn4@8c85907178da
Running 1m test @ http://localhost:5000/participants?limit=100
  1 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.86s   262.08ms   2.00s    94.58%
    Req/Sec     6.04      2.63    10.00     62.50%
  352 requests in 1.00m, 30.93MB read
  Socket errors: connect 0, read 0, write 0, timeout 186
Requests/sec:      5.86
Transfer/sec:    527.41KB
```

### Baseline - Biospecimens
```
wrk -t1 -c12 -d60s 'http://localhost:5000/biospecimens?limit=100'                                                                                                                  singhn4@8c85907178da
Running 1m test @ http://localhost:5000/biospecimens?limit=100
  1 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.11s   586.22ms   1.89s    66.67%
    Req/Sec     3.14      0.62     5.00     76.35%
  203 requests in 1.00m, 16.41MB read
  Socket errors: connect 0, read 0, write 0, timeout 197
Requests/sec:      3.38
Transfer/sec:    279.97KB
```
### Optimized - Biospecimens
```
wrk -t1 -c12 -d60s 'http://localhost:5000/biospecimens?limit=100'                                                                                                                  singhn4@8c85907178da
Running 1m test @ http://localhost:5000/biospecimens?limit=100
  1 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   903.60ms  120.18ms   1.46s    86.36%
    Req/Sec    13.10      5.03    20.00     61.85%
  792 requests in 1.00m, 64.04MB read
Requests/sec:     13.18
Transfer/sec:      1.07MB
```

### Baseline - Families
```
wrk -t1 -c12 -d60s 'http://localhost:5000/families?limit=100'                                                                                                                      singhn4@8c85907178da
Running 1m test @ http://localhost:5000/families?limit=100
  1 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.07s   580.48ms   1.83s    66.67%
    Req/Sec     3.03      0.60     5.00     76.65%
  197 requests in 1.00m, 7.57MB read
  Socket errors: connect 0, read 0, write 0, timeout 191
Requests/sec:      3.28
Transfer/sec:    129.09KB
```
### Optimized - Families
```
wrk -t1 -c12 -d60s 'http://localhost:5000/families?limit=100'                                                                                                                      singhn4@8c85907178da
Running 1m test @ http://localhost:5000/families?limit=100
  1 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   623.69ms   58.27ms 936.52ms   84.52%
    Req/Sec    18.93      4.93    30.00     76.61%
  1150 requests in 1.00m, 44.17MB read
Requests/sec:     19.14
Transfer/sec:    752.80KB
```

### Baseline - Studies
```
wrk -t1 -c12 -d60s 'http://localhost:5000/studies?limit=100'                                                                                                                       singhn4@8c85907178da
Running 1m test @ http://localhost:5000/studies?limit=100
  1 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.07s   631.75ms   1.98s    55.56%
    Req/Sec     3.85      1.55    10.00     88.94%
  226 requests in 1.00m, 25.26MB read
  Socket errors: connect 0, read 0, write 0, timeout 217
Requests/sec:      3.76
Transfer/sec:    430.60KB
```
### Optimized - Studies
```
wrk -t1 -c12 -d60s 'http://localhost:5000/studies?limit=100'                                                                                                                       singhn4@8c85907178da
Running 1m test @ http://localhost:5000/studies?limit=100
  1 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.00us    0.00us   0.00us     nan%
    Req/Sec     0.00      0.00     0.00    100.00%
  16 requests in 1.00m, 1.79MB read
  Socket errors: connect 0, read 0, write 0, timeout 16
Requests/sec:      0.27
Transfer/sec:     30.48KB
```

### Baseline - Investigators
```
wrk -t1 -c12 -d60s 'http://localhost:5000/investigators?limit=100'                                                                                                                 singhn4@8c85907178da
Running 1m test @ http://localhost:5000/investigators?limit=100
  1 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   186.07ms   18.51ms 315.97ms   79.86%
    Req/Sec    64.23      8.02    80.00     89.63%
  3853 requests in 1.00m, 7.17MB read
Requests/sec:     64.13
Transfer/sec:    122.24KB
```
### Optimized - Investigators
```
wrk -t1 -c12 -d60s 'http://localhost:5000/investigators?limit=100'                                                                                                                 singhn4@8c85907178da
Running 1m test @ http://localhost:5000/investigators?limit=100
  1 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   138.38ms   26.29ms 402.32ms   90.49%
    Req/Sec    86.81     14.09   110.00     79.66%
  5180 requests in 1.00m, 9.64MB read
Requests/sec:     86.32
Transfer/sec:    164.54KB
```

### Baseline - Sequencing Experiments
```
wrk -t1 -c12 -d60s 'http://localhost:5000/sequencing-experiments?limit=100'                                                                                                        singhn4@8c85907178da
Running 1m test @ http://localhost:5000/sequencing-experiments?limit=100
  1 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.07s   534.90ms   1.76s    66.67%
    Req/Sec     2.95      0.64     5.00     71.73%
  191 requests in 1.00m, 15.92MB read
  Socket errors: connect 0, read 0, write 0, timeout 185
Requests/sec:      3.18
Transfer/sec:    271.49KB
```
### Optimized - Sequencing Experiments
```
wrk -t1 -c12 -d60s 'http://localhost:5000/sequencing-experiments?limit=100'                                                                                                        singhn4@8c85907178da
Running 1m test @ http://localhost:5000/sequencing-experiments?limit=100
  1 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   775.19ms  112.90ms   1.38s    87.87%
    Req/Sec    15.19      5.03    20.00     58.41%
  923 requests in 1.00m, 76.93MB read
Requests/sec:     15.38
Transfer/sec:      1.28MB
```